Initial Commit
Textures are commonly used for everyday graphics implementations, such
as modeling,
rendering and display. As the bandwidth of hardware maintains its
current
rate of improvement, random access memory and speed continue to outpace
these
improvements, creating a bottle neck for accessing subsequent
information that is
needed for the next operations. Improving textures by making them
smaller to
push through the pipeline, or able to access more in a single operation
would be
able to overcome bandwidth limitations.